backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0

# LLM
openai>=1.40.0

# Database
aiosqlite>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Multimodal
edge-tts>=6.1.0           # TTS - free, no API key
pillow>=10.0.0            # Image processing
matplotlib>=3.8.0         # Chart generation

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9   # File upload support
httpx>=0.27.0
